requirements.txt
Flask==1.0.2
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Pillow==6.0.0
pytesseract==0.2.6
Werkzeug==0.15.2
wrapt==1.11.1
